Added html and css files
As noted in Canvas, the HTML file should contain a layout of different sections with the markup for each of the listed items above. This HTML file should contain a link to a CSS file. The CSS file should contain the styling for each of the different elements. HTML and CSS files were added to the repository as an example of a style guide.